Add caught exception as suppressed in SftpSession
The `org.springframework.integration.sftp.session.SftpSession#rename` method catches an instance of `SftpException`
but does not actually use it.

This makes is harder to troubleshoot issues as this exception cannot be accessed by the callers.

This commit adds this exception as "suppressed" to the exceptions that can be thrown by this method.